Moved/spaced-out code for readability and access.
