+++ frontend/model/contracts/group.js
@@ -15,7 +15,7 @@ import {
 import { paymentStatusType, paymentType, PAYMENT_COMPLETED } from './payments/index.js'
 import * as Errors from '../errors.js'
 import { merge, deepEqualJSONType, omit } from '~/frontend/utils/giLodash.js'
-import { currentMonthstamp, ISOStringToMonthstamp, compareMonthstamps } from '~/frontend/utils/time.js'
+import { currentMonthstamp, ISOStringToMonthstamp, compareMonthstamps, monthlyCycleStatsAtDate } from '~/frontend/utils/time.js'
 import { vueFetchInitKV } from '~/frontend/views/utils/misc.js'
 import groupIncomeDistribution from '~/frontend/utils/distribution/group-income-distribution.js'
 import currencies, { saferFloat } from '~/frontend/views/utils/currencies.js'
@@ -105,38 +105,6 @@ function initFetchMonthlyPayments ({ meta, state, getters }) {
   const monthlyPayments = vueFetchInitKV(state.paymentsByMonth, monthstamp, initPaymentMonth({ getters }))
   clearOldPayments({ state })
   return monthlyPayments
-}
-
-function addMonths (startDate, months) {
-  const thisMonth = startDate.getMonth()
-  startDate.setMonth(thisMonth + months)
-  if (startDate.getMonth() !== thisMonth + months && startDate.getMonth() !== 0) {
-    startDate.setDate(0)
-  }
-  return startDate
-}
-
-function monthlyCycleStatsAtDate ({ state, atDate }) {
-  const cycleNowDate = new Date(atDate)
-
-  const cycleStartDateInitial = new Date(state.distributionCycleStartDate)
-  const cycleEndDateInitial = addMonths(new Date(cycleStartDateInitial.toISOString()), 1)
-
-  let cycleStartDate = new Date(cycleStartDateInitial.toISOString())
-  let cycleEndDate = new Date(cycleEndDateInitial.toISOString())
-
-  let monthIteration = 0
-  while (cycleEndDate - cycleNowDate < 0) {
-    monthIteration++
-    cycleStartDate = addMonths(cycleStartDateInitial, monthIteration)
-    cycleEndDate = addMonths(cycleStartDateInitial, monthIteration + 1)
-  }
-
-  const cycleNow = (cycleNowDate - cycleStartDate) / (cycleEndDate - cycleStartDate) + monthIteration
-  const cycleStart = Math.floor(cycleNow)
-  const cycleEnd = cycleStart + 1
-
-  return { cycleNowDate, cycleStartDate, cycleEndDate, cycleNow, cycleStart, cycleEnd }
 }
 
 DefineContract({
@@ -279,9 +247,6 @@ DefineContract({
       //       bound to the UI in some location.
       return getters.groupSettings.mincomeCurrency && currencies[getters.groupSettings.mincomeCurrency].displayWithCurrency
     },
-    currentMonthCycleStats (state, getters) {
-      return monthlyCycleStatsAtDate({ state: getters.currentGroupState, getters, atDate: new Date().toISOString() })
-    },
     groupCreationDate (state, getters) {
       return getters.currentGroupState.distributionCycleStartDate
     },
@@ -412,7 +377,15 @@ DefineContract({
             const toUser = vueFetchInitKV(fromUser, data.toUser, [])
             toUser.push(data.paymentHash)
           }
-          state.distributionEvents.push({ type: 'paymentEvent', data: { from: meta.username, to: payment.data.toUser, amount: payment.data.amount, when: meta.createdDate, cycle: monthlyCycleStatsAtDate({ state, atDate: meta.createdDate }).cycleNow } })
+          state.distributionEvents.push({ type: 'paymentEvent',
+            data: {
+              from: meta.username,
+              to: payment.data.toUser,
+              amount: payment.data.amount,
+              when: meta.createdDate,
+              cycle: monthlyCycleStatsAtDate(state.groupCreationDate, meta.createdDate).cycleNow
+            }
+          })
           paymentMonth.lastAdjustedDistribution = groupIncomeDistribution({
             state, getters, updateMonthstamp, adjusted: true
           })
@@ -718,7 +691,14 @@ DefineContract({
         }
 
         const income = data.incomeDetailsType === 'incomeAmount' ? data.incomeAmount - state.settings.mincomeAmount : data.pledgeAmount
-        state.distributionEvents.push({ type: 'incomeDeclaredEvent', data: { name: meta.username, income, when: meta.createdDate, cycle: monthlyCycleStatsAtDate({ state, atDate: meta.createdDate }).cycleNow } })
+        state.distributionEvents.push({ type: 'incomeDeclaredEvent',
+          data: {
+            name: meta.username,
+            income,
+            when: meta.createdDate,
+            cycle: monthlyCycleStatsAtDate(state.groupCreationDate, meta.createdDate).cycleNow
+          }
+        })
       }
     },
     'gi.contracts/group/updateAllVotingRules': {
